test2

+++ deploy/scripts/init_dev.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
+
 cd /var/www/html/dev.mixcart.ru/
 export COMPOSER_HOME="$HOME/.composer";
 composer update
 php yii migrate --interactive=0
+if [ $? -gt 0 ]; then
+    exit 1
+fi
 
 sudo rm -rf /var/www/html/dev.mixcart.ru/frontend/web/assets/*
 sudo rm -rf /var/www/html/dev.mixcart.ru/backend/web/assets/*
